requirements.txt
# Web Framework & UI
streamlit>=1.29.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# NLP & Text Processing
PySastrawi>=1.2.0
scikit-learn>=1.3.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0

# Visualization
plotly>=5.18.0
